+++ libcorewar/sources/libft_printf/sources/core/conversions/u.c
@@ -6,7 +6,7 @@
 /*   By: afeuerst <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/31 08:50:48 by afeuerst          #+#    #+#             */
-/*   Updated: 2019/11/23 10:05:38 by afeuerst         ###   ########.fr       */
+/*   Updated: 2019/12/14 11:12:59 by afeuerst         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@ static const char	g_u_base[] = "0123456789";
 size_t				precalculate_u(struct s_printformat *const printformat, struct s_percent *const percent)
 {
 	percent->r1 = ft_printf_unumber_length(percent->data, 10);
-	if (percent->width && percent->width > percent->r1)
+	if (percent->width && (size_t)percent->width > percent->r1)
 	{
-		percent->r2 = percent->width - percent->r1;
+		percent->r2 = (size_t)percent->width - percent->r1;
 		return (percent->width);
 	}
 	else
